% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat2Phantasus_expressionset.R
\name{Seurat2Phantasus_expressionset}
\alias{Seurat2Phantasus_expressionset}
\title{A Seurat2Phantasus_expressionset Function
This function allows you to make pseudobulk expression set file}
\usage{
Seurat2Phantasus_expressionset(
  seuratobject,
  slot = "counts",
  assay = "RNA",
  sample_to_subset,
  cluster_to_subset
)
}
\arguments{
\item{seuratobject}{Your seuratobject}

\item{slot}{Slot: counts - raw, data - log2 Default is counts slot}

\item{assay}{Assay: RNA, SCT, etc Default is RNA slot}

\item{sample_to_subset}{sample metadata annotation}

\item{cluster_to_subset}{cluster metadata annotation}
}
\description{
A Seurat2Phantasus_expressionset Function
This function allows you to make pseudobulk expression set file
}
